refactor: exolix api
