+++ lib/services/exchange/exolix/api/dto/exolix_base_dto.dart
@@ -1,0 +1,6 @@
+abstract class ExolixBaseDto {
+  Map<String, dynamic> toMap();
+
+  @override
+  String toString() => toMap().toString();
+}
